print import covariates

+++ run.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 from src.model import Config, Model
 from src.data_layer import DataInRamInputLayer
 from src.utils import deco_print, deco_print_dict
-from src.utils_anlys import feature_ranking
+from src.utils_anlys import feature_ranking, feature_ranking_pair
 
 tf.flags.DEFINE_string('logdir', '', 'Path to save logs and checkpoints')
 tf.flags.DEFINE_string('mode', 'train', 'Mode: train/test/sens_anlys/sens_anlys_pair/sens_anlys_trio')
@@ -185,42 +185,45 @@ with tf.Session() as sess:
 		### real-valued pairs
 		num_feature_pairs = dl._covariate_count_float * (dl._covariate_count_float - 1) // 2
 		idx2pair = [(i+dl._covariate_count_int,j+dl._covariate_count_int) for i in range(dl._covariate_count_float-1) for j in range(i+1,dl._covariate_count_float)]
+		pair2idx = {idx2pair[l]:l for l in range(num_feature_pairs)}
 		###
 
-		pair2idx = {idx2pair[l]:l for l in range(num_feature_pairs)}
-		gradients = np.zeros(shape=(5, model._config.num_category, num_feature_pairs))
+		if not os.path.exists(os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.py')):
+			gradients = np.zeros(shape=(5, model._config.num_category, num_feature_pairs))
+			epoch_start = time.time()
+			cur_epoch_step = 0
+			sample_step = 0
+			for _, (x, y, info, x_cur) in enumerate(dl.iterate_one_epoch(model._config.batch_size, output_current_status=True)):
+				if sample_step != FLAGS.sample_size:
+					count += np.sum(x_cur, axis=0)
+					f, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x})
+					for i, j in idx2pair:
+						x_copy_1 = copy.deepcopy(x)
+						x_copy_2 = copy.deepcopy(x)
+						### 1 delta
+						x_copy_1[:,i] *= FLAGS.delta
+						f_i, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_1})
+						x_copy_2[:,j] *= FLAGS.delta
+						f_j, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_2})
+						### 2 delta
+						x_copy_1[:,j] *= FLAGS.delta
+						f_ij, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_1})
 
-		epoch_start = time.time()
-		cur_epoch_step = 0
-		sample_step = 0
-		for _, (x, y, info, x_cur) in enumerate(dl.iterate_one_epoch(model._config.batch_size, output_current_status=True)):
-			if sample_step != FLAGS.sample_size:
-				count += np.sum(x_cur, axis=0)
-				f, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x})
-				for i, j in idx2pair:
-					x_copy_1 = copy.deepcopy(x)
-					x_copy_2 = copy.deepcopy(x)
-					### 1 delta
-					x_copy_1[:,i] *= FLAGS.delta
-					f_i, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_1})
-					x_copy_2[:,j] *= FLAGS.delta
-					f_j, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_2})
-					### 2 delta
-					x_copy_1[:,j] *= FLAGS.delta
-					f_ij, = sess.run(fetches=[model._prob], feed_dict={model._x_placeholder:x_copy_1})
+						finite_diff = np.absolute(f_ij - f_i - f_j + f)
+						gradients[:,:,pair2idx[(i,j)]] += x_cur.T.dot(finite_diff)
+					sample_step += 1
+				if info['epoch_step'] != cur_epoch_step:
+					epoch_last = time.time() - epoch_start
+					time_est = epoch_last / (info['idx_file'] + 1) * info['num_file']
+					deco_print('Elapse / Estimate: %.2fs / %.2fs     ' %(epoch_last, time_est), end='\r')
+					cur_epoch_step = info['epoch_step']
+					sample_step = 0
+			gradients /= count[:, np.newaxis, np.newaxis]
+			deco_print('Saving Output in %s' %os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.npy'))
+			np.save(os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.npy'), gradients)
 
-					finite_diff = np.absolute(f_ij - f_i - f_j + f)
-					gradients[:,:,pair2idx[(i,j)]] += x_cur.T.dot(finite_diff)
-				sample_step += 1
-			if info['epoch_step'] != cur_epoch_step:
-				epoch_last = time.time() - epoch_start
-				time_est = epoch_last / (info['idx_file'] + 1) * info['num_file']
-				deco_print('Elapse / Estimate: %.2fs / %.2fs     ' %(epoch_last, time_est), end='\r')
-				cur_epoch_step = info['epoch_step']
-				sample_step = 0
-		gradients /= count[:, np.newaxis, np.newaxis]
-		deco_print('Saving Output in %s' %os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.npy'))
-		np.save(os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.npy'), gradients)
+		deco_print('Top 30: ')
+		feature_ranking_pair(FLAGS.logdir, dl._idx2covariate, idx2pair)
 		deco_print('Sensitivity Analysis (Pairs) Finished')
 
 	elif FLAGS.mode == 'sens_anlys_trio':
